pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nexus-terminal"
version = "2.0.0"
description = "NEXUS — Personal AI command center. Multi-LLM terminal with integrations."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Muhammad Rashid", email = "muhammadrashid@nexus.dev" },
]
keywords = ["ai", "terminal", "claude", "openai", "llm", "cli", "grok", "gemini"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
    "Topic :: Utilities",
]
dependencies = [
    "aiohttp>=3.9.0",
    "pywebview>=5.0",
    "psutil>=5.9.0",
]

[project.scripts]
nexus = "nexus_app.cli:main"

[project.urls]
Homepage = "https://github.com/muhammadrashid4587/nexus-terminal"
Repository = "https://github.com/muhammadrashid4587/nexus-terminal"
Issues = "https://github.com/muhammadrashid4587/nexus-terminal/issues"

[tool.hatch.build.targets.wheel]
packages = ["nexus_app"]
